Create mnhotsetup.bash
This script will set up a fresh Ubuntu VPS with the correct configuration file for running as an Odin masternode.
At the end you will be given a masternode key, which is what you will use to complete your masternode configuration on your local gui Odin wallet. 
This masternode key is NOT an Odin address.  Don't ever send any ODIN to it.  It's just a label that is used to link the hot and cold wallets together.